Add `SPLIT_PENALTY_COMPREHENSION` flag.
Also: Remove the coarse flag on the comprehension state calculation and
make the flags determine whether/which penalties are added.

I've set the new flag's value to 2100 for Google as that seems to be a
pretty strong bias towards keeping everything on the same line (but not
too strong that it upsets common cases).

80 might be too high for PEP8 but I think 0 penalty results in some
pretty lousy formatting.