Don't add space between '.' and print
To be honest, not sure we should ever be adding a space here, as
we can end up with class members that are keywords.

Also, I suspect that the existing logic is wrong, and should also be
checking rval, even if we don't want to add print to the checks.