Fix SPACES_BEFORE_COMMENT code block
The code in the section SPACES_BEFORE_COMMENT was using quote style rather than code block style, which has resulted in inconsistencies and several displaying problems. The code was meant to illustrate the number of spaces before a trailing comment, but quote style in Markdown does not implement it that way. This was corrected in the commit.