PCS-17: 1.7 Each line should contain less than 80 characters
According to the recommendation, each line should not exceed 80 characters
length. In order to enforce this recommendation we enclose long lines
in parentheses so that so that they will be correctly wrapped by
`Reformatter`. For instance

    if some_long_condition:
        pass

will be converted to

    if (some_long_condition):
        pass

then formatted with respect to COLUMN_LIMIT.

Currently only `if` and `while` statements are supported.

ling lines splitter