PCS-20: Rule 2.3 Common attribute comment
Actually, this changes affect not only attribute comments, but any
comments in input files. When INSERT_SPACES_AFTER_HASH_CHAR is on
missing spaces will be inserted in every commentary (excluding those
explicitly disabled via --lines). For example:

    '#some comment' -> '# some comment'

Note: there are two open issues in the mainstream project that would be
able to secure this rule but they do not seem to be resoved anytime soon:

    https://github.com/google/yapf/issues/139
    https://github.com/google/yapf/issues/512