yapf/yapflib/comment_formatter: correctly handle multiline comments
When a commentary contains several lines, they are spliced into a single
token in a single UnwrappedLine. The line's lineno is the lineno of the last
line in the commentary. Thus, we need can not rely on the line's lineno
only when check if a source file starts with shebang.