PCS-37: Rule 4.14 Do not use duplicate names
In python is possible to repeatedly define a class, function or method,
it this case the original object will be overwritten. While python allow
to redefine a function or class name, this is rarely a good idea and
might be a sign of incorrect code.

If WARN_REDEFINITION is on, for the example below a warning would be generated:

    1 def foo():
    2     pass
    3
    4 def foo(x):
    5     pass

    WARN 7: [filename: ..., line: 4]: 'foo' already defined here: lineno=1