PCS-41: Recommendation 4.7 The function code should be encapsulated
Top-level code should be encapsulated in functions or class. It is
recommended to check `if __name__ == '__main__` in scripts:

```
    #! /usr/bin/env python

    def main():
        pass

    if __name__ == '__main__':
        main()
```

This patch introduces CHECK_SCRIPT_CODE_ENCAPSULATION options. If this
option is enabled, then a warning will be generated whenever a script
(a source file with shebang) does not have such a check.