check the variable naming style only for lvalues
We should not check function arguments when a function is on the left
side of an assignment expression (e.g. `fn(arg) = value`). So is for the
function name itself. We also should not warn assignments for chained variable
names (except for `self.var = value`), because in this case variables are,
most likely, reassigned, not defined.

Before this patch the following expression would trigger a incorrect warning
at `None`:

    df.loc[slice(1, None)] = <something>