original spaces between lines are computet incorrectly in some cases
After `SpliceComments()` is applied we can not guarantee that the AST
tree can always be visited in the correct order (i.e. with respect to
line numbers). For instances, in the example below the comment will be
visited after `else` close.

    if condition:
        pass
    # comment
    else:
        pass

The corresponding (simplified) AST tree would look as following:
    if_stmt
        NAME if
        NAME condition
        COLON
        suite
            simple_stmt
                NAME pass
        NAME else
        COLON
        suite
            simple_stmt
                NAME pass
    simple_stmt
        COMMENT '# comment'