fix test affected by SAVE_INITIAL_INDENTS_FORMATTING
There is some rather weird logic describing whether a function is a
top-level one when it is not in the column 0 (which affects the number of
blank lines around the function definition). The presence of other statements
and comments determines the top-levelness of the function. This behaviour
has changed since we started storing the actual column value for comments,
when this value does not have to always be 0.

The issue is introduced by the following commit:
    keep original comments' indents when SAVE_ORIGINAL_INDENTS_FORMATTING=True

Failed tests:
    yapftests.reformatter_basic_test.BasicReformatterTest.testBlankLinesBeforeFunctionsNotInColumnZero
    yapftests.yapf_test.HorizontallyAlignedTrailingCommentsTest.testBlockIndentedFuncSuffix